server/requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community
langchain_experimental

#Grok
langchain-Groq

#youtube
youtube-transcript-api

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

# Environment Variable Management
python-dotenv

#flask
flask
Flask-Cors
#Chroma 
langchain_chroma

